// frontend/alanda/libs/common/src/lib/shared/api/models/project.ts
import { AlandaProcess } from './process';
import { ProjectState } from '../../../enums/projectState.enum';
import { AlandaMilestone } from './milestone';
import { AlandaSimplePhase } from './simplePhase';
import { AlandaHistoryLog } from './historyLog';
import { AlandaProperty } from './property';
import { AlandaProjectType } from './projectType';

export interface AlandaProject {
  guid?: number;
  version?: number;
  projectId?: string;
  tag?: string[];
  title?: string;
  status?: ProjectState;
  priority?: number;
  refObjectIdName?: string;
  refObjectType?: string;
  projectType?: string;
  projectTypeIdName?: string;
  subType?: string;
  dueDate?: string;
  pmcProjectType?: AlandaProjectType;
  subtype?: string;
  parents?: AlandaProject[];
  children?: AlandaProject[];
  processes?: AlandaProcess[];
  customerProjectId?: number;
  ownerId?: number;
  details?: string;
  guStatus?: string;
  comment?: string;
  risk?: number;
  refObjectId?: number;
  milestones?: AlandaMilestone[];
  phases?: AlandaSimplePhase[];
  history?: AlandaHistoryLog[];
  displayMetaInfo?: string;
  properties?: AlandaProperty[];
  propertiesMap?: Map<string, any>;
  milestonesMap?: Map<string, any>;
  createDate?: string;
  createUser?: number;
  ownerName?: string;
  updateDate?: string;
  updateUser?: number;
  parentIds?: number[];
  childrenIds?: number[];
  additionalInfo?: Map<string, any>;
  resultStatus?: string;
  resultComment?: string;
  authBase?: string;
  running?: boolean;
  humanReadableId?: string;
}